+++ gwcelery/conf/playground.py
@@ -6,3 +6,7 @@ sentry_environment = 'playground'
 """Record this `environment tag
 <https://docs.sentry.io/enriching-error-data/environments/>`)` in Sentry log
 messages."""
+
+mock_events_simulate_multiple_uploads = True
+"""If True, then upload each mock event several times in rapid succession with
+random jitter in order to simulate multiple pipeline uploads."""